+++ flake.nix
@@ -67,7 +67,7 @@
 
         ashBuild = ocamlPackages.buildDunePackage {
           pname = "ash";
-          version = "0.1.12";
+          version = "0.1.13";
           inherit src;
           duneVersion = "3";
 
@@ -136,7 +136,7 @@
 
         ash-command-pages = pkgs.stdenvNoCC.mkDerivation {
           pname = "ash-command-pages";
-          version = "0.1.12";
+          version = "0.1.13";
           dontUnpack = true;
           installPhase = ''
             runHook preInstall